Improved the front-end of the games.html view to show a leader board
Modified SalvoController to include score information for each GamePlayer in a Game.
Added getScore() to GamePlayer and Player to be able to recover the info for use in the GamePlayerDTO.
Modified the html, css and js files to show a leader board based on the score for each GamePlayer in a Game.
Made a correction in the CommandLineRunner (wrong player was being passed to a Score constructor)
Made an improvement to the game.js files (the getJSON method now logs a ser response status error message for anything other that status = 200).